feat: refactor adding readByExtension
