+++ src/types/Output.d.ts
@@ -1,6 +1,6 @@
 import { Spectrum1D, Spectrum2D, Molecule } from 'cheminfo-types';
 
-interface readZipOutput {
+interface Output {
   spectra: Array<Spectrum1D | Spectrum2D>;
   molecules: Array<Molecule>;
 }
